OneDrive/Documents/asx-swing-engine/requirements.txt
yfinance>=0.2.38
pandas>=2.0
numpy>=1.26
requests>=2.31
matplotlib>=3.8
mplfinance>=0.12.10b0
python-dotenv>=1.0
ib_insync>=0.9.86
streamlit>=1.35
plotly>=5.18
